- moved to repositories.exceptions package
